Update README.md
This project is a front-end clone of the Amazon homepage, designed to closely replicate the layout, structure, and design elements of the real Amazon.com interface. Built using HTML, CSS, and JavaScript, the goal of this clone is to improve practical UI/UX development skills and demonstrate the ability to create responsive, user-friendly e-commerce web pages.